requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dateparser==1.1.3
Flask==2.2.2
Flask-Mail==0.9.1
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
pymon==1.1
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
ta==0.10.2
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.1
